Updated Expense functions
I realized that if your budget name has either a space character or a
apostrophe character it would not create a table and throw an error. I
have remedied the situation by implementing a function that converts it
to a good SQL name for any function that relies on it.