Under Graduate Research Day Release
Added Icons.
Renamed project to Budget Buddy
Fixed a crash that occurred when you deleted a budget and then tried to
view a selected budget.
Created a Static Variable in Main Activity which holds the string for
the background color. This is so it is easy to revert colors as needed.
We can do this through the XML but ain't nobody got time for dat.